+++ covalent_dispatcher/_db/write_result_to_db.py
@@ -256,9 +256,7 @@ def transaction_insert_electrons_data(
 
     session.add(electron_row)
     session.flush()
-    electron_id = electron_row.id
-
-    return electron_id
+    return electron_row.id
 
 
 def insert_electrons_data(*args, **kwargs):
@@ -283,7 +281,7 @@ def transaction_insert_electron_dependency_data(
         dependency information of an electron
     """
 
-    # TODO - Update how we access the transport graph edges directly in favor of using some interface provied by the TransportGraph class.
+    # TODO - Update how we access the transport graph edges directly in favor of using some interface provided by the TransportGraph class.
     node_links = nx.readwrite.node_link_data(lattice.transport_graph._graph)["links"]
 
     electron_dependency_ids = []