+++ tests/covalent_dispatcher_tests/_core/execution_test.py
@@ -155,7 +155,7 @@ def test_get_task_inputs():
     result_object = Result(lattice=received_lattice, dispatch_id="asdf")
     tg = received_lattice.transport_graph
 
-    assert list(tg._graph.nodes) == [0, 1, 2, 3, 4, 5, 6, 7]
+    assert list(tg._graph.nodes) == list(range(17))
     tg.set_node_value(0, "output", ct.TransportableObject(1))
     tg.set_node_value(2, "output", ct.TransportableObject(2))
 
@@ -374,8 +374,8 @@ async def test_run_workflow_with_client_side_postprocess(test_db, mocker):
 
     result_object = await run_workflow(result_object)
     mock_unregister.assert_called_with(result_object.dispatch_id)
-    assert result_object.status == Result.PENDING_POSTPROCESSING
-    mock_run_abstract_task.assert_awaited_once()
+    assert result_object.status == Result.RUNNING
+    assert mock_run_abstract_task.call_count == 2
 
 
 @pytest.mark.asyncio
@@ -412,7 +412,7 @@ async def test_run_workflow_with_failed_postprocess(test_db, mocker):
     result_object = await run_workflow(result_object)
     mock_unregister.assert_called_with(result_object.dispatch_id)
 
-    assert result_object.status == Result.POSTPROCESSING_FAILED
+    assert result_object.status == Result.RUNNING
 
     result_object.lattice.workflow_function = ct.TransportableObject(failing_workflow)
     result_object.lattice.set_metadata("workflow_executor", "local")
@@ -420,8 +420,8 @@ async def test_run_workflow_with_failed_postprocess(test_db, mocker):
     result_object = await run_workflow(result_object)
     mock_unregister.assert_called_with(result_object.dispatch_id)
 
-    assert result_object.status == Result.POSTPROCESSING_FAILED
-    assert mock_run_abstract_task.call_count == 2
+    assert result_object.status == Result.RUNNING
+    assert mock_run_abstract_task.call_count == 4
 
 
 @pytest.mark.asyncio