+++ tests/covalent_dispatcher_tests/_db/update_test.py
@@ -23,11 +23,13 @@ import shutil
 from datetime import datetime as dt
 from datetime import timezone
 from pathlib import Path
+from unittest.mock import MagicMock
 
 import pytest
 
 import covalent as ct
 from covalent._results_manager.result import Result
+from covalent._shared_files.defaults import postprocess_prefix
 from covalent._workflow.lattice import Lattice as LatticeClass
 from covalent.executor import LocalExecutor
 from covalent_dispatcher._db import update, upsert
@@ -149,7 +151,7 @@ def test_update_node(test_db, result_1, mocker):
             == "test_sublattice"
         )
 
-        assert lattice_record.electron_num == 5
+        assert lattice_record.electron_num == 11
         assert lattice_record.completed_electron_num == 0
         assert lattice_record.updated_at is not None
     update._node(
@@ -175,7 +177,7 @@ def test_update_node(test_db, result_1, mocker):
         )
         assert result == 5
 
-        assert lattice_record.electron_num == 5
+        assert lattice_record.electron_num == 11
         assert lattice_record.completed_electron_num == 1
         assert lattice_record.updated_at is not None
 
@@ -279,7 +281,7 @@ def test_result_persist_workflow_1(test_db, result_1, mocker):
                 assert executor_data["attributes"] == le.__dict__
 
         # Check that there are the appropriate amount of electron dependency records
-        assert len(electron_dependency_rows) == 4
+        assert len(electron_dependency_rows) == 12
 
         # Update some node / lattice statuses
         cur_time = dt.now(timezone.utc)
@@ -287,7 +289,7 @@ def test_result_persist_workflow_1(test_db, result_1, mocker):
         result_1._status = "COMPLETED"
         result_1._result = ct.TransportableObject({"helo": 1, "world": 2})
 
-        for node_id in range(5):
+        for node_id in range(10):
             update._node(
                 result_1,
                 node_id=node_id,
@@ -318,17 +320,20 @@ def test_result_persist_workflow_1(test_db, result_1, mocker):
         assert result_1.result == result.get_deserialized()
 
         # Check that the electron records are as expected
-        for electron in electron_rows:
-            assert electron.status == "COMPLETED"
-            assert electron.parent_lattice_id == 1
-            assert (
-                electron.started_at.strftime("%Y-%m-%d %H:%M")
-                == electron.completed_at.strftime("%Y-%m-%d %H:%M")
-                == cur_time.strftime("%Y-%m-%d %H:%M")
-            )
-            assert Path(electron.storage_path) == Path(
-                f"{TEMP_RESULTS_DIR}/dispatch_1/node_{electron.transport_graph_node_id}"
-            )
+        for i, electron in enumerate(electron_rows):
+            if electron.name.startswith(postprocess_prefix):
+                assert electron.status == "NEW_OBJECT"
+            else:
+                assert electron.status == "COMPLETED"
+                assert electron.parent_lattice_id == 1
+                assert (
+                    electron.started_at.strftime("%Y-%m-%d %H:%M")
+                    == electron.completed_at.strftime("%Y-%m-%d %H:%M")
+                    == cur_time.strftime("%Y-%m-%d %H:%M")
+                )
+                assert Path(electron.storage_path) == Path(
+                    f"{TEMP_RESULTS_DIR}/dispatch_1/node_{electron.transport_graph_node_id}"
+                )
 
     # Tear down temporary results directory
     teardown_temp_results_dir(dispatch_id="dispatch_1")
@@ -343,18 +348,12 @@ def test_result_persist_subworkflow_1(test_db, result_1, result_2, mocker):
     update.persist(result_2, electron_id=1)
 
     with test_db.session() as session:
-        electron = session.query(Electron).where(Electron.id == 1).first()
+        session.query(Electron).where(Electron.id == 1).first()
         job_record = session.query(Job).where(Job.id == Electron.job_id).first()
-        job_record.cancel_requested = True
 
     # Query lattice / electron / electron dependency
     with test_db.session() as session:
         lattice_row = session.query(Lattice).where(Lattice.dispatch_id == "dispatch_2").first()
-        job_records = (
-            session.query(Electron, Job)
-            .where(Job.id == Electron.job_id, Electron.parent_lattice_id == lattice_row.id)
-            .all()
-        )
 
         # Check that lattice record is as expected
         assert lattice_row.dispatch_id == "dispatch_2"
@@ -367,9 +366,6 @@ def test_result_persist_subworkflow_1(test_db, result_1, result_2, mocker):
         assert lattice_row.electron_id == 1
         assert lattice_row.executor == "local"
         assert lattice_row.workflow_executor == "local"
-
-        for job in job_records:
-            assert job.cancel_requested is True
 
 
 def test_result_persist_rehydrate(test_db, result_1, mocker):
@@ -384,7 +380,8 @@ def test_result_persist_rehydrate(test_db, result_1, mocker):
         result_2 = _result_from(lattice_row)
 
     assert result_1.__dict__.keys() == result_2.__dict__.keys()
-    assert result_1.lattice.__dict__.keys() == result_2.lattice.__dict__.keys()
+    result_2.lattice._bound_electrons = {}
+    assert set(result_1.lattice.__dict__.keys()) == set(result_2.lattice.__dict__.keys())
     for key in result_1.lattice.__dict__.keys():
         if key == "transport_graph":
             continue
@@ -417,3 +414,41 @@ def test_lattice_persist(result_1):
 def test_transport_graph_persist(result_1):
     update.persist(result_1.lattice.transport_graph)
     assert result_1.lattice.transport_graph.dirty_nodes == []
+
+
+@pytest.mark.parametrize("node_name", [None, "mock_node_name", postprocess_prefix])
+def test_node(mocker, node_name):
+    """Test the _node method."""
+    electron_data_mock = mocker.patch("covalent_dispatcher._db.upsert.electron_data")
+    lattice_data_mock = mocker.patch("covalent_dispatcher._db.upsert.lattice_data")
+    mock_result = MagicMock()
+    update._node(
+        mock_result,
+        node_id=0,
+        node_name=node_name,
+        start_time="mock_time",
+        end_time="mock_time",
+        status="COMPLETED",
+        output="mock_output",
+    )
+    if node_name is None:
+        node_name = mock_result.lattice.transport_graph.get_node_value()
+    mock_result._update_node.assert_called_once_with(
+        node_id=0,
+        node_name=node_name,
+        start_time="mock_time",
+        end_time="mock_time",
+        status="COMPLETED",
+        output="mock_output",
+        error=None,
+        sub_dispatch_id=None,
+        sublattice_result=None,
+        stdout=None,
+        stderr=None,
+    )
+    if node_name.startswith(postprocess_prefix):
+        assert mock_result._result == "mock_output"
+        assert mock_result._status == "COMPLETED"
+    else:
+        assert mock_result._result != "mock_output"
+        assert mock_result._status != "COMPLETED"